1. etapa 2 - reorganizando archivos - refactorizando código - agregando gitignore y readme
