+++ .gitignore
@@ -1,0 +1,2 @@
+etapa2/models/__pycache__
+etapa2/tests/__pycache__
