+++ etapa2/README.md
@@ -1,0 +1,9 @@
+# Codo a Codo 
+## Full Stack Python
+### Trabajo Práctico Final
+### Etapa2
+
+1. Para testear las clases ejecutar 
+```python:
+python3 -m etapa2.tests.main-test
+```
