--- etapa2/tests/colores.py
@@ -1,9 +1,0 @@
-arrayColores = [
-    {"nombreColor":"rojo", "valorHexadec":"#f00"},
-    {"nombreColor":"verde", "valorHexadec":"#0f0"},
-    {"nombreColor":"azul", "valorHexadec":"#00f"},
-    {"nombreColor":"cyan", "valorHexadec":"#0ff"},
-    {"nombreColor":"magenta", "valorHexadec":"#f0f"},
-    {"nombreColor":"amarillo", "valorHexadec":"#ff0"},
-    {"nombreColor":"negro", "valorHexadec":"#000"}
-]
