+++ etapa2/tests/main-test.py
@@ -1,10 +1,10 @@
-from .tests.testProducto import testProducto
-from .tests.testInventario import testInventario
-from .tests.testCarrito import testCarrito
+from .testProducto import testProducto
+from .testInventario import testInventario
+from .testCarrito import testCarrito
 
 bandera = 1
 while bandera == 1:
-    opcion = int(input("1:Test 1 - 2:Test 2 - 3:Test 3 :"))
+    opcion = int(input("0: Exit - 1:Test 1 - 2:Test 2 - 3:Test 3 :"))
     if opcion == 1: 
         testProducto()
     if opcion == 2:
